% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dates.R
\name{impute_pdates}
\alias{impute_pdates}
\title{Impute start or end dates}
\usage{
impute_pdates(dates, ptype, input_sep = "-")
}
\arguments{
\item{dates}{a character vector of partial dates (which could also contain
full dates) in the format YYYY-MM-DD}

\item{ptype}{a string of either \code{"start"} or \code{"end"} indicating whether start
or end dates should be imputed, respectively}

\item{input_sep}{the character that separates date components in \code{dates},
default is \code{"-"}}
}
\value{
a date vector of imputed dates in the format YYYY-MM-DD
}
\description{
Partial dates should be the format UNKN-UN-UN or some combination of those
characters and numbers (ie 2017-UN-UN). Dates with no information as converted
to \code{NA}. For start dates, missing days are assumed to be the first of the
month while missing months are assumed to be January. For end dates, missing
days are assumed to be the last day of the month and missing months are
assumed to be December.
}
