% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dates.R
\name{reshape_adates}
\alias{reshape_adates}
\title{Reshape format of all dates (full and partial)}
\usage{
reshape_adates(dates)
}
\arguments{
\item{dates}{a character vector of dates}
}
\value{
a character vector of dates
}
\description{
Re-arranges full and partial dates in the general form of \code{"MM/DD/YYYY"} to
the ISO 8601 format (\code{"YYYY-MM-DD"}).
}
\details{
The date component separator in the input vector \code{dates} is irrelevant.
}
