% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dates.R
\name{trim_dates}
\alias{trim_dates}
\title{Trim unknown elements in partial dates}
\usage{
trim_dates(dates, input_sep = "-")
}
\arguments{
\item{dates}{a character vector of full or partial dates in the format
YYYY-MM-DD, for unknown dates it will be a variant of UNKN-UN-UN}

\item{sep}{the character that separates date components in the input vector
\code{dates}, default is \code{"-"}}
}
\value{
a character vector of trimmed dates
}
\description{
Trims \code{"2017-UN-UN"} to \code{"2017"} and trims \code{"2017-05-UN"} to \code{"2017-05"}.
Values of \code{"UNKN-UN-UN"} are converted to \code{NA}. Values where only
the month and day are known are converted to just the year, ie \code{"2017-UN-01"}
converts to \code{"2017"}. Full dates are not modified.
}
