% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sdtmdates-package.R
\docType{package}
\name{sdtmdates-package}
\alias{sdtmdates}
\alias{sdtmdates-package}
\title{sdtmdates: Transform and Impute Dates for SDTM}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Statistical programmers are required to transform data snapshots from an electronic data cut (EDC) into Clinical Data Interchange Standards Consortium (CDSIC) Study Data Tabulation Model (SDTM) compliant data sets per the guidelines in the SDTM Implementation Guide (SDTMIG). {sdtmdates} provides utility functions for statistical programmers to convert full and partial dates into SDTMIG compliant dates.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/skgithub14/sdtmdates}
  \item \url{https://skgithub14.github.io/sdtmdates/}
  \item Report bugs at \url{https://github.com/skgithub14/sdtmdates/issues}
}

}
\author{
\strong{Maintainer}: Stephen Knapp \email{stephen@knappconsultingllc.com} (\href{https://orcid.org/0000-0002-5101-4555}{ORCID})

}
\keyword{internal}
